requirements.txt
# AI Service Platform - Requirements
# Оптимизировано для Timeweb App Platform

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# HTTP
httpx==0.25.2
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0

# Telegram бот
python-telegram-bot==20.7

# Опционально (для расширенных возможностей)
# openai==1.3.7
# pillow==10.1.0
